Basic working bot completed

+++ bot.py
@@ -1,4 +1,5 @@
-# bot.py
+#!/usr/bin/env python3
+
 import os
 import mysql.connector
 import discord
@@ -47,7 +48,7 @@ async def ping(ctx):
 	await ctx.send('Yes, Master {}, I am listening!'.format(author))
  
 @bot.command(name='signin', help='Sign In to the event')
-async def signin(ctx, *, character):
+async def signin(ctx, *, player: str=""):
 	# TODO: fix - signing in to no/deleted event
 	# TODO: set display name to be the character name
 	author = ctx.message.author.display_name
@@ -65,13 +66,13 @@ async def signin(ctx, *, character):
 		await ctx.send('I\'m sorry, Master {}, I was unable to check your name on the member list.'.format(nameid))	
 	
 	member_id = 0
-	if count == 0:
-
-		await ctx.send('TEST')
+	if count == 0 and player == "":
+		await ctx.send('I\'m sorry, Master, please let me know your Out Of Character name for the membership records')
+	elif count == 0:
 		#Add to member table
 		try:
-			sql = "INSERT INTO members (name, wordpress_id) VALUES (%s, %s)"
-			val = (nameid, "")
+			sql = "INSERT INTO members (name, playername, wordpress_id) VALUES (%s, %s, %s)"
+			val = (nameid, player, "")
 			mycursor.execute(sql, val)
 			mydb.commit()
 			member_id = mycursor.lastrowid
@@ -93,9 +94,8 @@ async def signin(ctx, *, character):
 			print(e)
 			await ctx.send('I\'m sorry, Master, I was unable to find your membership number.')
 
-
 	if member_id > 0:
-
+		
 		# Is there an event today and has it started?
 		isevent = 0
 		try:
@@ -103,30 +103,36 @@ async def signin(ctx, *, character):
 			mycursor.execute(sql)
 			events = mycursor.fetchall()
 			#await ctx.send("Thank you, Master, the {} event is in progress.".format(events[0][1]))
-			isevent = 1
+			isevent = len(events)
 		except Exception as e:
 			print(e)
 			await ctx.send('I\'m sorry, Master, I could not read the diary to check for an event today.')
 			isevent = -1
-
+			
 		# Yes? Sign-in
-		if isevent == 1:
+		if isevent > 0:
+		
+			eventid = events[0][0]
 
 			# Have you already signed in?
 			signedin = 0
 			try:
-				sql = "select count(member_id) from attendance where member_id = '{}'".format(member_id)
-				mycursor.execute(sql)
+				sql = "select count(member_id) from attendance, events where events.id = attendance.event_id and attendance.member_id = %s and events.id = '%s'"
+				val = (member_id, eventid)
+				mycursor.execute(sql, val)
 				out = mycursor.fetchall()
 				signedin = out[0][0]
 			except Exception as e:
 				print(e)
 				await ctx.send('I\'m sorry, Master, I\'m unable to confirm if you have already signed in.')
 
+
 			if signedin == 0:
 				try:
-					sql = "INSERT INTO attendance (member_id, event_id, charactername) VALUES (%s, %s, %s)"
-					val = (member_id, events[0][0], character)
+					sql = "INSERT INTO attendance (member_id, event_id, displayname) VALUES (%s, %s, %s)"
+					#await ctx.send(sql)
+					#await ctx.send(events)
+					val = (member_id, events[0][0], author)
 					mycursor.execute(sql, val)
 					mydb.commit()
 					await ctx.send('Thank you Master {}, I have recorded your attendance'.format(author))
@@ -146,10 +152,10 @@ async def signin(ctx, *, character):
 		await ctx.send('I\'m sorry Master {}, your membership has not been confirmed'.format(author))
 	
 
-@signin.error
-async def signin_error(ctx, error):
-	if isinstance(error, commands.MissingRequiredArgument):
-		await ctx.send("I'm sorry Master, Who should I say is attending the event?")
+# @signin.error
+# async def signin_error(ctx, error):
+	# if isinstance(error, commands.MissingRequiredArgument):
+		# await ctx.send("I'm sorry Master, Who should I say is attending the event?")
 
 
 @bot.command(name='new', help='Schedule a new LARP event', usage='<name> dd/mm/yyyy [08:00pm]')
@@ -158,26 +164,51 @@ async def new(ctx, name: str, date: str, time: str='08:00pm'):
 	author = ctx.message.author.display_name
 	server = ctx.message.guild.name
 	date_time = '{} {}'.format(date, time)
-	# TODO: check that event is not in the past
-	# TODO: check event doesn't already exist
-	#....
 	# add to database
+	dateok = 0
 	try:
 		event_date = datetime.strptime(date_time, '%d/%m/%Y %I:%M%p')
-		sql = "INSERT INTO events (name, server, eventdate) VALUES (%s, %s, %s)"
-		val = (name, server, event_date)
-		mycursor.execute(sql, val)
-		mydb.commit()
-		await ctx.send('Thank you Master {}, I have recorded the {} event in the diary for {}'.format(author, name, event_date))
+		dateok = 1
 	except Exception as e:
-		await ctx.send('I\'m sorry, Master, I could not complete your command')
+		await ctx.send('I\'m sorry, Master, the format of the date and optional time is DD/MM/YYYY [08:00pm]')
 		print(e)
+	
+	# check that event is not in the past
+	present = datetime.now()
+	if event_date.date() < present.date():
+		await ctx.send('I\'m sorry, Master, event date you have specified is in the past.')
+		dateok = 0
+	
+	# check event doesn't already exist
+	isduplicate = 1
+	try:
+		sql = "SELECT COUNT(id) FROM events WHERE name = '{}'".format(name)
+		mycursor.execute(sql)
+		out = mycursor.fetchall()
+		isduplicate = out[0][0]
+	except Exception as e:
+		await ctx.send('I\'m sorry, Master, I could not check if an event already exists with this name.')
+		print(e)
+	
+	if isduplicate:
+		await ctx.send('I\'m sorry, Master, I already have an event with this name.')
+	
+	if dateok and isduplicate == 0:
+		try:
+			sql = "INSERT INTO events (name, server, eventdate) VALUES (%s, %s, %s)"
+			val = (name, server, event_date)
+			mycursor.execute(sql, val)
+			mydb.commit()
+			await ctx.send('Thank you Master {}, I have recorded the {} event in the diary for {}'.format(author, name, event_date))
+		except Exception as e:
+			await ctx.send('I\'m sorry, Master, I could not complete your command')
+			print(e)
 
 @new.error
 async def new_error(ctx, error):
 	if isinstance(error, commands.MissingRequiredArgument):
 		await ctx.send("I'm sorry Master, I need more information. Can you tell me the {} of the event?".format(error.param.name))
-
+	
 
 
 @bot.command(name='list', usage='[all | next | <event>]', brief='list game events and attendance', help='list game events and attendance.\n\nall\t: all events and attendance\nnext\t: next event\n<name>\t: list of attendees')
@@ -215,24 +246,39 @@ async def list(ctx, *, action: str='next'):
 			events = mycursor.fetchall()
 			await ctx.send("Master {}, the next event is {} and it takes place on the {}".format(author, events[0][1], events[0][3]))
 		else:
-			# list users who attended
-			sql = """SELECT attendance.charactername, events.eventdate
-				FROM 
-					events,
-					attendance,
-					members
-				WHERE 
-					events.server = '{}'
-					AND events.name = '{}'
-					AND events.id = attendance.event_id
-					AND members.member_id = attendance.member_id
-				ORDER BY attendance.charactername""".format(guild, action)
+			# TODO: account for when no one has signed in
+			sql = """SELECT COUNT(attendance.id)
+				FROM
+					attendance, events
+				WHERE
+					attendance.event_id = events.id
+					AND events.name = '{}'""".format(action)
 			mycursor.execute(sql)
-			events = mycursor.fetchall()
-			headers = ['Name']
-			rows = [[e[0]] for e in events]
-			table = tabulate(rows, headers)
-			await ctx.send('```Attendance for the ' + action + ' event on the {}:\n\n'.format(events[0][1]) + table + '```')
+			countall = mycursor.fetchall()
+			count = countall[0][0]
+			
+			if count > 0:
+				# list users who attended
+				sql = """SELECT attendance.displayname, members.playername, events.eventdate
+					FROM 
+						events,
+						attendance,
+						members
+					WHERE 
+						events.server = '{}'
+						AND events.name = '{}'
+						AND events.id = attendance.event_id
+						AND members.member_id = attendance.member_id
+					ORDER BY attendance.displayname""".format(guild, action)
+				#await ctx.send(sql)
+				mycursor.execute(sql)
+				events = mycursor.fetchall()
+				headers = ['Character', 'Player']
+				rows = [[e[0], e[1]] for e in events]
+				table = tabulate(rows, headers)
+				await ctx.send('```Attendance for the ' + action + ' event on the {}:\n\n'.format(events[0][2]) + table + '```')
+			else:
+				await ctx.send('```Attendance for the ' + action + ' event:\n\nNo attendees```')
 	except Exception as e:
 		await ctx.send("I'm sorry Master, I am unable to provide the event details you requested")
 		print(e)
@@ -243,16 +289,39 @@ async def list(ctx, *, action: str='next'):
 async def delete(ctx, name: str):
 	author = ctx.message.author.display_name
 	guild = ctx.message.guild.name
-	# TODO: check if event exists before trying to delete it
-	# TODO: delete associated attendance as well
+	# check if event exists before trying to delete it
+	event_id = 0
 	try:
-		sql = "DELETE FROM `events` WHERE name = '{}'".format(name)
+		sql = "select id from events where name = '{}'".format(name)
 		mycursor.execute(sql)
-		mydb.commit()
-		await ctx.send('Of course Master {}, I have removed the {} event from the schedule.'.format(author, name))
+		events = mycursor.fetchall()
+		event_id = events[0][0]
+		#await ctx.send('I have found events ID {} with that name.'.format(event_id))
 	except Exception as e:
-		await ctx.send("I'm sorry Master, I am unable to remove the event from the calendar.")
 		print(e)
+		await ctx.send('I\'m sorry, Master {}, I was unable to check for that event.'.format(nameid))	
+
+	if event_id > 0:
+		# delete associated attendance as well
+		try:
+			sql = "DELETE FROM `attendance` WHERE event_id = '{}'".format(event_id)
+			mycursor.execute(sql)
+			mydb.commit()
+			#await ctx.send('I have cleared the attendance list for the event. {} attendees have been removed'.format(mycursor.rowcount))
+		except Exception as e:
+			await ctx.send("I'm sorry Master, I am unable to clear the attendance list.")
+			print(e)
+		
+		try:
+			sql = "DELETE FROM `events` WHERE name = '{}'".format(name)
+			mycursor.execute(sql)
+			mydb.commit()
+			await ctx.send('Of course Master {}, I have removed the {} event from the schedule.'.format(author, name))
+		except Exception as e:
+			await ctx.send("I'm sorry Master, I am unable to remove the event from the calendar.")
+			print(e)
+	else:
+		await ctx.send('I\'m sorry, Master {}, the {} event does not exist'.format(author, name))	
 
 @delete.error
 async def delete_error(ctx, error):
